voronoi: fixed shared data init bug, small changes

+++ src/files/files.h
@@ -10,9 +10,9 @@
 #define MAX_NUM_FILES 10000
 
 enum error_code add_files_sorted(const char *path, char *names[MAX_NUM_FILES],
-                                 size_t *len);
+                                 int *len);
 bool something_to_do(char *const names[MAX_NUM_FILES], const char *source,
-                     const char *destination, size_t len,
+                     const char *destination, int len,
                      long *latest_source_file);
 enum error_code process_file(const char *name, const char *source,
                              const char *destination,
